Further edited the File Check and Delete functionality. Previously, if interrupted, some files would remain when removed from log. A new runnable has been created to delete all files with same date. uploadFile method in UploadService has been changed to a boolean to enable exceptions to be handled. moveFile method has been moved out of uploadFile (as re-uploaded files throw exceptions). Integer values for locating date / id substrings in filename are now stored rather than calculated.
Returned GPS Sample to a long. Short was too small. Some huge recordings may be due to short cycling and becoming negative. Completelyavoidable.

Release to NTT.

+++ app/src/main/java/uk/ac/nottingham/eaxtp1/CradleRideLogger/IMUService.java
@@ -54,7 +54,7 @@ public class IMUService extends Service implements SensorEventListener {
     static boolean heldWithMagnets;
 
     private long sampleID;
-    private short prevSample;
+    private long prevSample;
 
     String sID, sX, sY, sZ, sampleTime, toQueue,
             sGyX = "", sGyY = "", sGyZ = "", sE = "", sN = "", sD = "", sAmp = "";
